I created a BusStop Class , and a JsonFileReader.Class. The BusStop Class is used for storing BusStop information such as name, location and the busses that stop at the specific bus stop, I use it  in an array to create the Bus Stop Markers
The JsonFileReader class can only read the UCSC_Westside_Bus_Stop.json file , if you try to input any other json file it will break the app.

I added a new icon for the bus stops. I changed it to a bench with a stop sign instead of a bus.

Also NetworkActivity was crashing because pref = null so i changed it to pref = "